feat: 데이터베이스에 저장된 정보로 chessboard 생성 구현
