+++ docker/db/mysql/init/initfortest.sql
@@ -28,4 +28,20 @@ VALUES ('a', 1, 'ROOK', 'WHITE'),
        ('e', 2, 'PAWN', 'WHITE'),
        ('f', 2, 'PAWN', 'WHITE'),
        ('g', 2, 'PAWN', 'WHITE'),
-       ('h', 2, 'PAWN', 'WHITE');
+       ('h', 2, 'PAWN', 'WHITE'),
+       ('a', 8, 'ROOK', 'BLACK'),
+       ('b', 8, 'KNIGHT', 'BLACK'),
+       ('c', 8, 'BISHOP', 'BLACK'),
+       ('d', 8, 'QUEEN', 'BLACK'),
+       ('e', 8, 'KING', 'BLACK'),
+       ('f', 8, 'BISHOP', 'BLACK'),
+       ('g', 8, 'KNIGHT', 'BLACK'),
+       ('h', 8, 'ROOK', 'BLACK'),
+       ('a', 7, 'PAWN', 'BLACK'),
+       ('b', 7, 'PAWN', 'BLACK'),
+       ('c', 7, 'PAWN', 'BLACK'),
+       ('d', 7, 'PAWN', 'BLACK'),
+       ('e', 7, 'PAWN', 'BLACK'),
+       ('f', 7, 'PAWN', 'BLACK'),
+       ('g', 7, 'PAWN', 'BLACK'),
+       ('h', 7, 'PAWN', 'BLACK');
