+++ src/test/java/chess/dao/ChessGameDaoTest.java
@@ -10,35 +10,18 @@ import chess.domain.position.File;
 import chess.domain.position.Position;
 import chess.domain.position.Rank;
 import chess.dto.ChessGameComponentDto;
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.IOException;
 import java.sql.Connection;
-import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.List;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class ChessGameDaoTest {
+class ChessGameDaoTest extends DaoTest {
     ChessGameDao chessGameDao;
 
     @BeforeEach
-    void setUp() {
+    void initializeChessGameDao() {
         chessGameDao = new ChessGameDao();
-        try {
-            executeInitScript();
-        } catch (IOException | SQLException e) {
-            throw new RuntimeException(e);
-        }
-        System.setProperty("TEST_ENV", "true");
-    }
-
-    @AfterEach
-    void tearDown() {
-        System.setProperty("TEST_ENV", "false");
     }
 
     @DisplayName("데이터베이스 연결이 되었는지 확인한다.")
@@ -53,7 +36,7 @@ class ChessGameDaoTest {
     void findAll() {
         List<ChessGameComponentDto> dtos = chessGameDao.findAll();
 
-        assertThat(dtos.size()).isEqualTo(16);
+        assertThat(dtos.size()).isEqualTo(32);
     }
 
     @DisplayName("데이터베이스에 데이터를 저장한다.")
@@ -64,7 +47,7 @@ class ChessGameDaoTest {
 
         chessGameDao.save(chessGameComponentDto);
 
-        assertThat(chessGameDao.findAll().size()).isEqualTo(17);
+        assertThat(chessGameDao.findAll().size()).isEqualTo(33);
     }
 
     @DisplayName("데이터베이스에서 position에 해당되는 piece를 찾아온다.")
@@ -107,27 +90,5 @@ class ChessGameDaoTest {
         Piece sourcePiece = chessGameDao.findPieceByPosition(target);
 
         assertThat(sourcePiece).isNull();
-    }
-
-    private void executeInitScript() throws IOException, SQLException {
-        try (BufferedReader reader = new BufferedReader(new FileReader("docker/db/mysql/init/initfortest.sql"));
-             Connection connection = chessGameDao.getConnection();
-             Statement statement = connection.createStatement()) {
-            String line;
-            StringBuilder scriptContent = new StringBuilder();
-            while ((line = reader.readLine()) != null) {
-                handleScriptLine(line, scriptContent, statement);
-            }
-        }
-    }
-
-    private void handleScriptLine(String line, StringBuilder scriptContent, Statement statement) throws SQLException {
-        if (!line.trim().isEmpty() && !line.trim().startsWith("#")) {
-            scriptContent.append(line).append("\n");
-            if (line.trim().endsWith(";")) {
-                statement.execute(scriptContent.toString());
-                scriptContent.setLength(0);
-            }
-        }
     }
 }
